[breaking] Remove the default Console transport from the default Logger. [doc] Update several examples in examples/*.js
