--- examples/silly-levels.js
@@ -1,21 +1,0 @@
-var winston = require('../lib/winston');
-
-var myCustomLevels = {
-    levels: {
-        foo: 0,
-        bar: 1,
-        baz: 2,
-        foobar: 3
-    },
-    colors: {
-        foo: 'blue',
-        bar: 'green',
-        baz: 'yellow',
-        foobar: 'red'
-    }
-};
-
-var customLevelLogger = winston.createLogger({ levels: myCustomLevels.levels });
-customLevelLogger.foobar('some foobar level-ed message');
-customLevelLogger.foobar('some foobar level-ed message');
-customLevelLogger.foobar('some foobar level-ed message');
