+++ lib/winston.js
@@ -69,14 +69,7 @@ winston.loggers = new winston.Container();
 //     winston.log('info', 'some message');
 //     winston.error('some error');
 //
-var defaultLogger = winston.createLogger({
-  //
-  // TODO (indexzero): Remove this transport and display a common
-  // warning if logs are attempted to be written with
-  // no transports.
-  //
-  transports: [new winston.transports.Console()]
-});
+var defaultLogger = winston.createLogger();
 
 //
 // Pass through the target methods onto `winston.
