+++ lib/winston/logger.js
@@ -200,7 +200,14 @@ Logger.prototype._transform = function _transform(info, enc, callback) {
   // quite confusing when using custom levels.
   //
   if (!this.levels[info[LEVEL]] && this.levels[info[LEVEL]] !== 0) {
-    console.error('Unknown logger level: %s', info[LEVEL]);
+    console.error('[winston] Unknown logger level: %s', info[LEVEL]);
+  }
+
+  //
+  // Remark: not sure if we should simply error here.
+  //
+  if (!this._readableState.pipes) {
+    console.error('[winston] Attempt to write logs with no transports %j', info)
   }
 
   //
